pyproject.toml
[project]
name = "aoc-toolkit"
version = "0.1.0"
description = "Advent of Code toolkit to get, run and submit problems"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.7.0",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.1",
    "lxml>=5.3.0",
    "markdownify>=0.14.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
